+++ src/Components/PDFViewer/PDFViewer.vue
@@ -22,7 +22,7 @@
 -->
 
 <template>
-	<iframe :src="pdf">
+	<iframe :src="url">
 		{{ t("libresign", "This iframe is not supported in your browser.") }}
 	</iframe>
 </template>
@@ -38,15 +38,13 @@ export default {
 			default: '',
 		},
 	},
-
-	data() {
-		return {
-			pdf: this.url,
-		}
-	},
 }
 </script>
 
 <style lang="scss" scoped>
-@import './styles';
+iframe {
+	overflow: hidden;
+	height: 100%;
+	width: 100%;
+}
 </style>